Refactor Autocomplete code to BasicAutocomplete
