+++ examples/basic-app/model/model.json
@@ -61,7 +61,7 @@
                       "label": {
                         "field": "title"
                       },
-                      "ent": "vxg/task"
+                      "ent": "vxg/talk"
                     }
                   },
                   {
@@ -72,7 +72,7 @@
                       "label": {
                         "field": "title"
                       },
-                      "ent": "vxg/task"
+                      "ent": "vxg/talk"
                     }
                   }
                 ]
